tools/monitoring/requirements.txt
GitPython==3.1.3
jinja2==2.11.2
